requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
Flask==3.0.3
ffmpeg-python==0.2.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.1
MarkupSafe==2.1.5
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
psycopg2-binary>=2.9.0
